MyCourseManagement를 ProfessorCourseManagement로 수정하였습니다
